+++ 3D-Graph-Learn/EngineWindow.cpp
@@ -129,8 +129,8 @@ bool EngineWindow::HandleError(std::wstring errorMessage) {
 }
 
 void EngineWindow::ComputeElapsedTime() {
-	float time = highResClockRatio * std::chrono::high_resolution_clock::now().time_since_epoch().count();
-	m_elapsedTime = time - m_time;
+	double time = highResClockRatio * std::chrono::high_resolution_clock::now().time_since_epoch().count();
+	m_elapsedTime = static_cast<float>(time - m_time);
 	m_time = time;
 }
